+++ bazel/core.bzl
@@ -182,9 +182,6 @@ MLN_CORE_SOURCE = [
     "src/mbgl/map/zoom_history.hpp",
     "src/mbgl/math/log2.cpp",
     "src/mbgl/platform/settings.cpp",
-    "src/mbgl/programs/attributes.hpp",
-    "src/mbgl/programs/segment.hpp",
-    "src/mbgl/programs/uniforms.hpp",
     "src/mbgl/renderer/backend_scope.cpp",
     "src/mbgl/renderer/bucket.hpp",
     "src/mbgl/renderer/bucket_parameters.cpp",
@@ -292,7 +289,10 @@ MLN_CORE_SOURCE = [
     "src/mbgl/renderer/transition_parameters.hpp",
     "src/mbgl/renderer/update_parameters.hpp",
     "src/mbgl/renderer/upload_parameters.hpp",
+    "src/mbgl/shaders/attributes.hpp",
     "src/mbgl/shaders/program_parameters.cpp",
+    "src/mbgl/shaders/segment.hpp",
+    "src/mbgl/shaders/uniforms.hpp",
     "src/mbgl/sprite/sprite_loader.cpp",
     "src/mbgl/sprite/sprite_loader.hpp",
     "src/mbgl/sprite/sprite_loader_observer.hpp",
@@ -848,6 +848,7 @@ MLN_OPENGL_SOURCE = [
     "src/mbgl/gl/vertex_buffer_resource.cpp",
     "src/mbgl/gl/vertex_buffer_resource.hpp",
     "src/mbgl/platform/gl_functions.cpp",
+    "src/mbgl/shaders/gl/legacy/clipping_mask_program.cpp",
     "src/mbgl/shaders/gl/legacy/clipping_mask_program.hpp",
     "src/mbgl/shaders/gl/legacy/program_base.hpp",
     "src/mbgl/shaders/gl/legacy/program.hpp",
