+++ src/mbgl/gfx/drawable_builder_impl.hpp
@@ -4,7 +4,7 @@
 #include <mbgl/gfx/index_vector.hpp>
 #include <mbgl/gfx/vertex_attribute.hpp>
 #include <mbgl/gfx/vertex_vector.hpp>
-#include <mbgl/programs/segment.hpp>
+#include <mbgl/shaders/segment.hpp>
 #include <mbgl/gfx/drawable_builder.hpp>
 #include <mbgl/gfx/polyline_generator.hpp>
 #include <mbgl/shaders/widevector_ubo.hpp>
