+++ src/mbgl/renderer/buckets/circle_bucket.hpp
@@ -6,7 +6,7 @@
 #include <mbgl/tile/geometry_tile_data.hpp>
 #include <mbgl/gfx/vertex_buffer.hpp>
 #include <mbgl/gfx/index_buffer.hpp>
-#include <mbgl/programs/segment.hpp>
+#include <mbgl/shaders/segment.hpp>
 #include <mbgl/style/layers/circle_layer_properties.hpp>
 
 namespace mbgl {
