+++ src/mbgl/renderer/buckets/hillshade_bucket.hpp
@@ -4,7 +4,7 @@
 #include <mbgl/renderer/paint_property_binder.hpp>
 #include <mbgl/gfx/vertex_buffer.hpp>
 #include <mbgl/geometry/dem_data.hpp>
-#include <mbgl/programs/segment.hpp>
+#include <mbgl/shaders/segment.hpp>
 #include <mbgl/renderer/bucket.hpp>
 #include <mbgl/renderer/tile_mask.hpp>
 #include <mbgl/style/layers/hillshade_layer_properties.hpp>
