+++ src/mbgl/renderer/layers/render_background_layer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <mbgl/programs/segment.hpp>
+#include <mbgl/shaders/segment.hpp>
 #include <mbgl/renderer/render_layer.hpp>
 #include <mbgl/style/layers/background_layer_impl.hpp>
 #include <mbgl/style/layers/background_layer_properties.hpp>
