+++ src/mbgl/renderer/paint_property_binder.hpp
@@ -5,7 +5,7 @@
 #include <mbgl/gfx/uniform.hpp>
 #include <mbgl/gfx/upload_pass.hpp>
 #include <mbgl/layout/pattern_layout.hpp>
-#include <mbgl/programs/attributes.hpp>
+#include <mbgl/shaders/attributes.hpp>
 #include <mbgl/renderer/cross_faded_property_evaluator.hpp>
 #include <mbgl/renderer/paint_property_statistics.hpp>
 #include <mbgl/renderer/possibly_evaluated_property_value.hpp>
