+++ src/mbgl/shaders/gl/legacy/clipping_mask_program.cpp
@@ -1,0 +1,3 @@
+#include <mbgl/shaders/gl/legacy/clipping_mask_program.hpp>
+
+namespace mbgl {} // namespace mbgl
