+++ src/mbgl/shaders/gl/legacy/clipping_mask_program.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <mbgl/shaders/gl/legacy/program.hpp>
-#include <mbgl/programs/attributes.hpp>
-#include <mbgl/programs/uniforms.hpp>
+#include <mbgl/shaders/attributes.hpp>
+#include <mbgl/shaders/uniforms.hpp>
 #include <mbgl/style/properties.hpp>
 
 namespace mbgl {
